requirements.txt
Django==4.1.9
django-cors-headers==4.1.0
djangorestframework==3.14.0
python-dateutil
mysqlclient
